feat(#832): decimal strtod32/64/128, fenv rounding/exception plumbing, constant folding
Phase 2 of #402: strtod32/64/128 (FFI-wrapper lowering, no new opcode,
following the #828 <decimal_math.h> precedent); fesetround() now has real
effect on decimal arithmetic and fetestexcept() observes decimal invalid/
divbyzero/overflow/underflow/inexact via a CCCC_DEC_ENV_DYNAMIC/_STATIC
split; and compile-time constant folding of decimal arithmetic in static/
global initializers plus the reverse decimal->binary-float/decimal->int
cast directions.

quantize/samequantum/quantexp/quantum (also named in #832) were already
delivered by #828 -- verified via the existing test suite, no new work.

Fixed an independently-verified pre-existing bug along the way: the host FP
environment could already be dirty (FE_UNDERFLOW) before any decimal code
ran, because tokenize.c's convert_pp_number scans every floating/decimal
literal via a host strtold() call whose side effect wasn't cleared. cc_run()
now resets the host FP environment exactly once before the compiled guest
program starts running.

Deferred to a follow-up ticket: decimal comparisons directly in an integer
constant expression, BID_ROUNDING_TIES_AWAY (no <fenv.h> equivalent), and
per-thread rounding semantics.

+++ tests/test_decimal_ice_error.c
@@ -1,11 +1,16 @@
 // EXPECT_COMPILE_ERROR
-// C23 (#402): decimal constant folding is not implemented in phase 1 (see
-// the follow-up ticket), so a _Decimal value cannot appear in an integer
-// constant expression (array bound, case label, bit-field width,
-// _Static_assert, enumerator) -- not even through a cast. Must produce a
-// clean diagnostic, never a silent 0.
+// C23: a _Decimal value whose *own* type is decimal cannot appear directly
+// in an integer constant expression (array bound, case label, bit-field
+// width, _Static_assert, enumerator) -- must produce a clean diagnostic,
+// never a silent 0. This is still rejected in both build configurations
+// after #832: `(int)1.5dd` (a decimal-to-integer *cast*, the immediate
+// operand case ISO C actually permits in an ICE) now folds correctly --
+// see tests/test_decimal_ice_fold.c -- but a bare, uncast decimal literal
+// used directly as an integer constant expression is not something #832
+// added folding for (there is no destination width to fold *to* without a
+// cast), so it remains a diagnostic.
 
-int arr[(int)1.5dd];
+int arr[1.5dd];
 
 int main(void) {
     return sizeof(arr);
